refactor(search): rename `pages` to `components`
