+++ snippets/src/components/search/register-shortcut.ts
