+++ snippets/src/components/search/search-modal-event.ts
