+++ snippets/src/components/search/search-modal/useSearchBrowser.tsx
