+++ snippets/src/components/search/search-modal/useShortcut.tsx
