+++ snippets/src/components/search/search-placeholder/main.ts
