+++ snippets/src/components/search/search-placeholder/search-placeholder.css
