+++ snippets/vite.config.ts
@@ -13,11 +13,11 @@ export default defineConfig({
     outDir: resolve(__dirname, '../static/js'),
     rollupOptions: {
       input: {
-        "404": resolve(__dirname, 'src/pages/404/index.html'),
-        "post": resolve(__dirname, 'src/pages/post/index.html'),
-        "search-modal": resolve(__dirname, 'src/pages/search/search-modal/index.html'),
-        "search-placeholder": resolve(__dirname, 'src/pages/search/search-placeholder/index.html'),
-        "commander": resolve(__dirname, 'src/pages/search/commander/index.html'),
+        "404": resolve(__dirname, 'src/components/404/index.html'),
+        "post": resolve(__dirname, 'src/components/post/index.html'),
+        "search-modal": resolve(__dirname, 'src/components/search/search-modal/index.html'),
+        "search-placeholder": resolve(__dirname, 'src/components/search/search-placeholder/index.html'),
+        "commander": resolve(__dirname, 'src/components/search/commander/index.html'),
       },
       output: {
         entryFileNames: `[name].js`,
